requirements.txt
pyyaml
requests
xmltodict
dict2xml
tabulate
bs4
tracking_url
dateparser
